Added abstract class BaseModelMethods, moved methods to this and added add method
